Rename package for HeroesOfMagic
