+++ src/main/java/faang/school/godbless/HeroesOfMagic/Creature.java
@@ -1,4 +1,4 @@
-package faang.school.godbless.task_1;
+package faang.school.godbless.HeroesOfMagic;
 
 import lombok.Getter;
