+++ src/main/java/faang/school/godbless/HeroesOfMagic/Swordman.java
@@ -1,4 +1,4 @@
-package faang.school.godbless.task_1;
+package faang.school.godbless.HeroesOfMagic;
 
 public class Swordman extends Creature{
     public Swordman(int quantity) {
